+++ stdlib/os.h
@@ -1,0 +1,20 @@
+#pragma once
+#include <stdint.h>
+#include <stddef.h>
+#include <stdio.h>
+#include <stdlib.h>
+#include <string.h>
+#include <math.h>
+#include <setjmp.h>
+#define TC_ALLOC(type, count) ((type *)calloc((count), sizeof(type)))
+#define TC_LENOF(x) (sizeof(x) / sizeof((x)[0]))
+#define TC_FAT_LENOF(x) ((x).len)
+
+
+int32_t os(char *cmd);
+
+
+int32_t os(char *cmd) {
+    
+    return system(cmd);
+}